Move movefile fetch responsability to dedicated class
