+++ lib/wordmove/movefile/doctor.rb
@@ -1,0 +1,4 @@
+module Wordmove
+  class MovefileDoctor
+  end
+end
