+++ spec/deployer/base_spec.rb
@@ -45,99 +45,17 @@ describe Wordmove::Deployer::Base do
     end
   end
 
-  context ".fetch_movefile" do
-    TMPDIR = "/tmp/wordmove".freeze
-
-    let(:path) { File.join(TMPDIR, 'movefile.yml') }
-    let(:yaml) { "name: Waldo\njob: Hider" }
+  context ".fecth_movefile" do
+    let(:movefile) { double('movefile') }
 
     before do
-      FileUtils.mkdir(TMPDIR)
-      allow(described_class).to receive(:current_dir).and_return(TMPDIR)
-      allow(described_class).to receive(:logger).and_return(double('logger').as_null_object)
+      allow(Wordmove::Movefile).to receive(:new).and_return(movefile)
     end
 
-    after do
-      FileUtils.rm_rf(TMPDIR)
-    end
+    it "calls Wordmove::Movefile .fetch method" do
+      expect(movefile).to receive(:fetch)
 
-    context "when Movefile is missing" do
-      it 'raises an exception' do
-        expect { described_class.fetch_movefile }.to raise_error(Wordmove::MovefileNotFound)
-      end
-    end
-
-    context "when Movefile is present" do
-      before do
-        File.open(path, 'w') { |f| f.write(yaml) }
-      end
-
-      it 'finds a Movefile in current dir' do
-        result = described_class.fetch_movefile
-        expect(result['name']).to eq('Waldo')
-        expect(result['job']).to eq('Hider')
-      end
-
-      context "when movefile has no extensions" do
-        let(:path) { File.join(TMPDIR, 'movefile') }
-
-        it 'finds it aswell' do
-          result = described_class.fetch_movefile
-          expect(result['name']).to eq('Waldo')
-          expect(result['job']).to eq('Hider')
-        end
-      end
-
-      context "when Movefile has no extensions and has first capital" do
-        let(:path) { File.join(TMPDIR, 'Movefile') }
-
-        it 'finds it aswell' do
-          result = described_class.fetch_movefile
-          expect(result['name']).to eq('Waldo')
-          expect(result['job']).to eq('Hider')
-        end
-      end
-
-      context "when movefile.yaml has long extension" do
-        let(:path) { File.join(TMPDIR, 'movefile.yaml') }
-
-        it 'finds it aswell' do
-          result = described_class.fetch_movefile
-          expect(result['name']).to eq('Waldo')
-          expect(result['job']).to eq('Hider')
-        end
-      end
-
-      context "directories traversal" do
-        before do
-          @test_dir = File.join(TMPDIR, "test")
-          FileUtils.mkdir(@test_dir)
-          allow(described_class).to receive(:current_dir).and_return(@test_dir)
-        end
-
-        it 'goes up through the directory tree and finds it' do
-          result = described_class.fetch_movefile
-          expect(result['name']).to eq('Waldo')
-          expect(result['job']).to eq('Hider')
-        end
-
-        context 'Movefile not found, met root node' do
-          it 'raises an exception' do
-            allow(described_class).to receive(:current_dir).and_return('/tmp')
-            expect { described_class.fetch_movefile }.to raise_error(Wordmove::MovefileNotFound)
-          end
-        end
-
-        context 'Movefile not found, found wp-config.php' do
-          before do
-            FileUtils.touch(File.join(@test_dir, "wp-config.php"))
-          end
-
-          it 'raises an exception' do
-            expect { described_class.fetch_movefile }.to raise_error(Wordmove::MovefileNotFound)
-          end
-        end
-      end
+      described_class.fetch_movefile
     end
   end
